gradio/requirements.txt
gradio>=4.21.0
requests

# PostgreSQL driver
psycopg2-binary

# OpenAI & LLM
openai
langchain
langgraph

langchain-community
langchain-openai
wikipedia

# Data processing & visualization
pandas
matplotlib
plotly
pillow
